+++ modules/task_3/zlobin_g_shortest_paths_moore_algorithm/moore_algorithm.h
@@ -5,8 +5,8 @@
 #define MAX_VALUE 20
 #define EDGE_PROB 0.5
 
-void fillGraphWithRandomValues(int* graph_to_fill, size_t size = 10);
-void getSequentialMooreAlgorithm(int* graph, size_t size, int* shortest_ways, size_t begin_vertex = 0);
-void getParallelMooreAlgorithm(int* graph, size_t size, int* shortest_ways, size_t begin_vertex = 0);
+void fillGraphWithRandomValues(int* graph_to_fill, int size = 10);
+void getSequentialMooreAlgorithm(int* graph, int size, int* shortest_ways, int begin_vertex = 0);
+void getParallelMooreAlgorithm(int* graph, int size, int* shortest_ways, int begin_vertex = 0);
 
 #endif  // MODULES_TASK_3_ZLOBIN_G_SHORTEST_PATHS_MOORE_ALGORITHM_MOORE_ALGORITHM_H_
